refactored code inline with other examples -fixing review comments 1. The self.can_move check is redundant here, because it's already done in self.move 2. model.agents is the new API, and is simpler. 3. The whole function can be simplified to return np.mean(model.agents.get("price")). 4. The intermediate variables are not necessary, because the docstring is already sufficiently descriptive. 5. added tests for slope
