+++ examples/hotelling_law/requirements.txt
@@ -2,3 +2,4 @@ mesa
 pandas
 matplotlib
 numpy
+scipy
