+++ examples/hotelling_law/visualization.py
